Material is now fully incorporated into medium. But there is a a circular dependency that prevents the material to be included in the medium_specific file, which is ugly
